Mandelbrot cache WIP

+++ ClusterFrac/source/Host.h
@@ -105,7 +105,7 @@ namespace cf
 		*/
 		inline int getClientsCount();
 
-		inline int getTasksCount() { std::unique_lock<std::mutex> lock(taskQueueMutex); return taskQueue.size(); };
+		inline int getTasksCount() { std::unique_lock<std::mutex> lock(taskQueueMutex); return (int)taskQueue.size(); };
 
 		bool checkAvailableResult(unsigned __int64 taskID);
